config route for dashboard & refactor project structure
