+++ src/java/actions/syllabus/GetSyllabusByID.java
@@ -10,12 +10,10 @@ import dao.CLODao;
 import dao.MaterialDao;
 import dao.PreRequisiteDao;
 import dao.SessionDao;
-import dao.SubjectDao;
 import dao.SyllabusDao;
 import entities.CourseLearningObjective;
 import entities.Material;
 import entities.Session;
-import entities.Subject;
 import entities.Syllabus;
 import exceptions.NotFoundException;
 import java.io.IOException;
