+++ src/java/controllers/Auth/AuthController.java
@@ -1,12 +1,12 @@
 /*
-     * Click nbfs://nbhost/SystemFileSystem/Templates/Licenses/license-default.txt to change this license
-     * Click nbfs://nbhost/SystemFileSystem/Templates/JSP_Servlet/Servlet.java to edit this template
+ * To change this license header, choose License Headers in Project Properties.
+ * To change this template file, choose Tools | Templates
+ * and open the template in the editor.
  */
-package controllers;
+package controllers.auth;
 
 import actions.auth.Login;
 import actions.auth.Logout;
-import actions.auth.UpdatePassword;
 import java.io.IOException;
 import java.io.PrintWriter;
 import java.util.HashMap;
@@ -20,16 +20,15 @@ import actions.Action;
 
 /**
  *
- * @author quocb
+ * @author Admin
  */
-@WebServlet(name = "Auth2", urlPatterns = {"/auth/*"})
-public class AuthPassWordController extends HttpServlet {
+@WebServlet(name = "UserController", urlPatterns = {"/login", "/logout"})
+public class AuthController extends HttpServlet {
 
     private Map<String, Action> actionMap = new HashMap<>();
-
     public void init() {
-        actionMap.put("/auth/reset-password", new UpdatePassword());
-        actionMap.put("/auth/update-password", new UpdatePassword());
+        actionMap.put("/login", new Login());
+        actionMap.put("/logout", new Logout());
     }
 
     /**
@@ -49,10 +48,10 @@ public class AuthPassWordController extends HttpServlet {
             out.println("<!DOCTYPE html>");
             out.println("<html>");
             out.println("<head>");
-            out.println("<title>Servlet NewServlet</title>");
+            out.println("<title>Servlet UserController</title>");
             out.println("</head>");
             out.println("<body>");
-            out.println("<h1>Servlet Auth2 at " + request.getContextPath() + "</h1>");
+            out.println("<h1>Servlet UserController at " + request.getContextPath() + "</h1>");
             out.println("</body>");
             out.println("</html>");
         }
@@ -70,9 +69,7 @@ public class AuthPassWordController extends HttpServlet {
     @Override
     protected void doGet(HttpServletRequest request, HttpServletResponse response)
             throws ServletException, IOException {
-        String path = request.getServletPath() + request.getPathInfo();
-        Action action = actionMap.get(path);
-        System.out.println("path " + path);
+        Action action = actionMap.get(request.getServletPath());
         if (action != null) {
             action.doGet(request, response);
         } else {
@@ -92,10 +89,8 @@ public class AuthPassWordController extends HttpServlet {
     @Override
     protected void doPost(HttpServletRequest request, HttpServletResponse response)
             throws ServletException, IOException {
-        System.out.println(request.getContextPath());
-        String path = request.getServletPath() + request.getPathInfo();
-        Action action = actionMap.get(path);
-        System.out.println("path " + path);
+        System.out.println("path " + request.getServletPath());
+        Action action = actionMap.get(request.getServletPath());
         if (action != null) {
             action.doPost(request, response);
         } else {
