+++ src/java/controllers/dashboard/DashboardController.java
@@ -2,10 +2,11 @@
  * Click nbfs://nbhost/SystemFileSystem/Templates/Licenses/license-default.txt to change this license
  * Click nbfs://nbhost/SystemFileSystem/Templates/JSP_Servlet/Servlet.java to edit this template
  */
-package controllers;
+package controllers.dashboard;
 
-import actions.auth.Login;
-import actions.auth.Logout;
+import actions.Action;
+import actions.curriculumns.GetCurriculumnByID;
+import actions.dashboard.ViewDashboard;
 import java.io.IOException;
 import java.io.PrintWriter;
 import java.util.HashMap;
@@ -15,20 +16,20 @@ import javax.servlet.annotation.WebServlet;
 import javax.servlet.http.HttpServlet;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
-import actions.Action;
-import actions.curriculumns.GetCurriculumnByID;
 
 /**
  *
  * @author quocb
  */
-@WebServlet(name = "CurriculumController", urlPatterns = {"/curriculums/*"})
-public class CurriculumController extends HttpServlet {
+@WebServlet(name = "DashboardController", urlPatterns = {"/dashboard/*"})
+public class DashboardController extends HttpServlet {
 
-    private final Map<String, Action> actionMap = new HashMap<>();
+    private final Map<String, HttpServlet> controllerMap = new HashMap<>();
+    private final Action dashboardAction = new ViewDashboard();
 
+    @Override
     public void init() {
-        actionMap.put("/curriculums", new GetCurriculumnByID());
+        controllerMap.put("curriculums", new CurriculumDashboard());
     }
 
     /**
@@ -43,18 +44,7 @@ public class CurriculumController extends HttpServlet {
     protected void processRequest(HttpServletRequest request, HttpServletResponse response)
             throws ServletException, IOException {
         response.setContentType("text/html;charset=UTF-8");
-        try ( PrintWriter out = response.getWriter()) {
-            /* TODO output your page here. You may use following sample code. */
-            out.println("<!DOCTYPE html>");
-            out.println("<html>");
-            out.println("<head>");
-            out.println("<title>Servlet CurriculumController</title>");
-            out.println("</head>");
-            out.println("<body>");
-            out.println("<h1>Servlet CurriculumController at " + request.getContextPath() + "</h1>");
-            out.println("</body>");
-            out.println("</html>");
-        }
+
     }
 
     // <editor-fold defaultstate="collapsed" desc="HttpServlet methods. Click on the + sign on the left to edit the code.">
@@ -69,12 +59,14 @@ public class CurriculumController extends HttpServlet {
     @Override
     protected void doGet(HttpServletRequest request, HttpServletResponse response)
             throws ServletException, IOException {
-        Action action = actionMap.get(request.getServletPath());
-        if (action != null) {
-            action.doGet(request, response);
+        String path = request.getPathInfo().split("/")[1];
+        HttpServlet servlet = controllerMap.get(path);
+
+        if (servlet != null) {
+            servlet.init();
+            servlet.service(request, response);
         } else {
-            // Handle error if the path is not supported
-            System.out.println("Not found action");
+            dashboardAction.doGet(request, response);
         }
     }
 
@@ -89,12 +81,14 @@ public class CurriculumController extends HttpServlet {
     @Override
     protected void doPost(HttpServletRequest request, HttpServletResponse response)
             throws ServletException, IOException {
-        Action action = actionMap.get(request.getServletPath());
-        if (action != null) {
-            action.doPost(request, response);
+        String path = request.getPathInfo().split("/")[1];
+        HttpServlet servlet = controllerMap.get(path);
+
+        if (servlet != null) {
+            servlet.init();
+            servlet.service(request, response);
         } else {
-            // Handle error if the path is not supported
-            System.out.println("Not found action");
+            dashboardAction.doPost(request, response);
         }
     }
